fix(quantization): upcast input in quantized-layer forwards for fp16/bf16 inference (ISS-018)
GPTQ/AWQ/SmoothQuant/PTQ layer forwards materialise fp32 dequantized
weights but passed the input through to F.linear un-upcast, so a half/
bfloat16 model (the HF loader and serving default) crashed with
'RuntimeError: mat1 and mat2 must have the same dtype'. The GPTQ
docstring already promised the upcast; it was never implemented.

Upcast the input to fp32 before the linear call in _gptq_layer,
_awq_layer, _smooth_layer and ptq so fp16/bf16 inputs work and the
output stays fp32 as documented. Adds a dtype regression test per layer
(each verified to fail pre-fix with the dtype RuntimeError).

+++ src/llm/quantization/_awq_layer.py
@@ -107,4 +107,6 @@ class AWQQuantizedLinear(nn.Module):
             scales_expanded = scales.to(torch.float32).repeat_interleave(gs, dim=1)
             w_fp = w_int_signed * scales_expanded
 
-        return torch.nn.functional.linear(x, w_fp, self.bias)
+        # Weights are materialised in fp32; upcast the input so fp16/bf16
+        # model inference works.  Output is fp32 (matches the docstring).
+        return torch.nn.functional.linear(x.to(torch.float32), w_fp, self.bias)
